requirements.txt
accelerate==1.4.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.47.2
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
audioread==3.0.1
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.20.0
azure-mgmt-compute==34.0.0
azure-mgmt-core==1.5.0
azure-mgmt-network==28.1.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
borb==2.1.25
boto3==1.37.0
botocore==1.37.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cli_exit_tools==1.2.7
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.0
cryptography==44.0.1
cssselect==1.2.0
cycler==0.12.1
Cython==3.1.0
dashscope==1.22.1
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.9
distro==1.9.0
docker==7.1.0
docker-pycreds==0.4.0
easyocr==1.7.2
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fabric==3.2.2
Farama-Notifications==0.0.4
fastdtw==0.3.4
filelock==3.17.0
Flask==3.0.3
fonttools==4.56.0
formulas==1.2.8
frozenlist==1.5.0
fsspec==2025.2.0
func_timeout==4.3.5
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
greenlet==3.1.1
groq==0.18.0
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
gymnasium==0.28.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
ImageHash==4.3.2
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
invoke==2.2.0
isodate==0.7.2
itsdangerous==2.2.0
jax-jumpy==1.0.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
lib-detect-testenv==2.0.8
librosa==0.10.2.post1
llvmlite==0.43.0
loguru==0.7.3
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib==3.7.5
MouseInfo==0.1.3
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
msgpack==1.1.0
mss==10.0.0
multidict==6.1.0
multiprocess==0.70.17
mutagen==1.47.0
networkx==3.2.1
ninja==1.11.1.3
numba==0.60.0
numpy==1.24.4
numpy-financial==1.0.0
oauth2client==4.1.3
odfpy==1.4.1
openai==1.64.0
opencv-python==4.8.1.78
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.0.0
platformdirs==4.3.6
playwright==1.50.0
pooch==1.8.2
portalocker==2.10.1
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
psutil==5.9.8
pyacoustid==1.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyDrive==1.3.1
pyee==12.1.1
pygame==2.6.1
PyGetWindow==0.0.9
PyJWT==2.10.1
PyMsgBox==1.0.9
PyMuPDF==1.25.3
PyNaCl==1.5.0
pynput==1.7.7
pyobjc-core==11.0
pyobjc-framework-ApplicationServices==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-CoreText==11.0
pyobjc-framework-Quartz==11.0
pyparsing==3.2.1
pypdf==5.3.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-barcode==0.15.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
pytweening==1.2.0
pytz==2025.1
PyWavelets==1.6.0
PyYAML==6.0.2
qrcode==8.0
RapidFuzz==3.12.1
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
rubicon-objc==0.5.0
s3transfer==0.11.2
safetensors==0.5.2
schedula==1.5.55
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
screeninfo==0.8.1
sentry-sdk==2.22.0
setproctitle==1.3.5
shapely==2.0.7
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2024.8.30
tiktoken==0.9.0
tokenizers==0.15.2
torch==2.5.1
torchvision==0.20.1
tqdm==4.65.2
transformers==4.35.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.20
wandb==0.19.7
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wrapt_timeout_decorator==1.5.1
XlsxWriter==3.2.2
xmltodict==0.14.2
yarl==1.18.3
zipp==3.21.0
